Fix location permissions warning issue
Fixed an issue where a location permissions warning alert controller was presented in all cases for a fresh install of the app.
The problem was that the app did not check if the location permissions were requested before, so even if the user was never presented with a prompt the warning case would trigger - meaning users would always be shown the permissions denied warning no matter what they chose.

The fix is to rely on the `CLLocationManagerDelegate` call back of `locationManagerDidChangeAuthorization(_ :)`
which is called both at initialization time of `CLLocationManager` and when permissions are changed.
The check at `appDidBecomeActive` time has been removed because the aforementioned delegate call accomplishes the same thing.